Add demo OrNull method for boolean only
This is for #186 / #282 and shows one approach only for boolean
(if this looks fine, we'd cut-and-paste for the other getters).

I used Optional here because it seems likely to be the way
of the future and produces a safer API.

Because of Optional, this API would be commented out if someone
does a Java 7 branch.