Reimplement ResolveMemos with a ridiculous hand-rolled hash table
This is intended to band-aid #330 with a hash table that sucks,
but ought to be much faster to copy than the stock Java one.
The stock Java hash table rehashes everything every time, while
this hash table can often just copy an array. That said, I didn't
benchmark it or anything, it may well also be super slow.

This is more or less a troll commit to get someone to do better.